update custom mapper docs with proper method signature
Add the `device` parameter to the method signature in the custom mapper documentation as well its explanation.

Add note about having to set the custom mapper in `settings.tres` in the custom mapper section itself. I know its mentioned earlier, but I thought it'd be nice to mention it in the section itself.